test: add coverage for relay_schema.py and ConfigData model
